Added Tile.push_tile_below method.
Tile.push_tile_below is an integral method for the way the grid works
in this game.  Although there are some optimizations to be made, it
does just what it says on the tin: The calling Tile attempts to push
the Tile below it down by one row, and that Tile attempts to do the
same, all the way down until the bottom Tile is checked.  If that tile
is at rock bottom, then it will return False and the top Tile cannot
push any of the other Tiles down.  Otherwise, all Tiles will drop one
row down.