Some minor fixes and changes.
I removed the Field.__getitem__ function, as I arrived at the belief
that I didn't gain anything by using Field[row][column] as a shorthand
for Field.cells[row][column].  Removing this leaves me with just one way
to reference a cell, as nature intended.

I also had to redo the collision detection, which is a lot more
complicated in Tritetris due to blocks being more than one cell in size.
It's a lot simpler in WWY, but I still had to rewrite it.