Added letters to tiles, functionality to wildcard tiles.
Tiles now display with their letters on them.

Also, using the 'z' key, when you get a wildcard tile, you can now run
through the current game's alphabet and pick a letter.  This will still
need to be fleshed out, but I was in proof-of-concept mode this morning
and just wanted to get something tangible on here.