Fixed a bug in the wildcard logic.
When a wildcard tile is dropped, if the player has not chosen a
letter yet, then the game chooses one at random. However, the
game was still randomly assigning a letter even after the player
chose a letter for the tile. This is because, in active_tile_has_landed,
I was only checking if the tile was a wildcard. I was not checking
if a letter had been assigned. This has been fixed, and the random
assignment now only happens when the tile is blank when it lands.